Add bad-whitespace disable to travis config for examples.
Whitespace in examples is explicit and makes them more readable with vertical alignment of pins.  Adding disable comments to the code will result in confusion (are they required to make the example work? no but not obvious to inexperienced users).  Disable this globally for all example like the invalid name and docstring checks.